postgres_to_es/requirements.txt
psycopg-binary==3.1.9
psycopg==3.1.9
pydantic==1.10.9
python-dotenv==1.0.0